Refactoring and adding a metadataForm.getControl method
